Move update_where to sqlite.query_helpers
